requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
transformers==4.35.2
torch==2.1.1
playwright==1.40.0
jinja2==3.1.2
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
